PIM-3232 Clean up router
