+++ src/Pim/Bundle/EnrichBundle/Resources/config/requirejs.yml
@@ -99,6 +99,7 @@ config:
     paths:
         pim/app:                            pimenrich/js/app
         pim/router:                         pimenrich/js/router
+        pim/route-matcher:                  pimenrich/js/route-matcher
         pim/page-title:                     pimenrich/js/page-title
 
         # Controllers
